Fix null-pointer dereference in soc_create_pcm()
When soc_create_pcm() finds all four components it needs, it displays a status
message.  This message references the local variables 'codec_dai' and
'cpu_dai', but these variables are initialized only when the codec DAI and the
cpu DAI are the last to be registered.  This patch changes the code to use the
stored values instead of the local variables.

Signed-off-by: Timur Tabi <timur@freescale.com>
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>