Update Freescale MPC8610 codec driver naming
Change the Freescale MPC8610 HPCD codec and fabric drivers to use the codec
name and id, instead of just the name, to match codec instances.  Previously,
the codec and fabric drivers would create a name string based on the
I2C bus and address.  Instead, keep the name as "cs4270" and combine the
bus and address into a unique number.

Signed-off-by: Timur Tabi <timur@freescale.com>
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>